refactored again folder structure
