+++ www/src/components/App.js
@@ -3,10 +3,10 @@ import Header from './Header'
 import Router from './Router'
 import './App.css'
 
-const App = (props) => (
+const App = () => (
   <div>
-    <Header {...props} />
-    <Router {...props} />
+    <Header />
+    <Router />
   </div>
 )
 export default App
