--- www/src/services/index.js
@@ -1,3 +1,0 @@
-import './logger'
-import './router'
-import './auth'
