+++ www/src/store/forms/index.js
@@ -1,0 +1,2 @@
+import './reducer'
+
