+++ www/src/store/index.js
@@ -2,3 +2,4 @@
 import './auth'
 import './router'
 import './forms'
+import './logger'
