+++ www/src/store/logger/index.js
@@ -1,0 +1,2 @@
+
+import './services'
