+++ www/src/store/router/index.js
@@ -1,0 +1,2 @@
+import './reducer'
+import './services'
